# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Future Jaano - AI-powered spiritual guidance platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kundli:
    post:
      operationId: createKundli
      tags: [kundli]
      summary: Generate Kundli (birth chart) analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KundliInput"
      responses:
        "201":
          description: Generated Kundli
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KundliReport"
  /kundli/{id}:
    get:
      operationId: getKundli
      tags: [kundli]
      summary: Get a previously generated Kundli
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kundli report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KundliReport"
...
  /vastu/analyze:
    post:
      operationId: analyzeVastu
      tags: [vastu]
      summary: Analyze a photo for Vastu Shastra insights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VastuInput"
      responses:
        "201":
          description: Vastu analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VastuReport"
  /vastu/{id}:
    get:
      operationId: getVastuReport
      tags: [vastu]
      summary: Get a Vastu analysis report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vastu report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VastuReport"

  # PALM READING
  /palm/analyze:
    post:
      operationId: analyzePalm
      tags: [palm-reading]
      summary: Analyze a palm image for reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageAnalysisInput"
      responses:
        "201":
          description: Palm reading result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PalmReport"
  /palm/{id}:
    get:
      operationId: getPalmReport
      tags: [palm-reading]
      summary: Get a palm reading report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Palm reading report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PalmReport"

  # FACE READING
  /face/analyze:
    post:
      operationId: analyzeFace
      tags: [face-reading]
      summary: Analyze a face image for reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageAnalysisInput"
      responses:
        "201":
          description: Face reading result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaceReport"
  /face/{id}:
    get:
      operationId: getFaceReport
      tags: [face-reading]
      summary: Get a face reading report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Face reading report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaceReport"

  # NUMEROLOGY
...
components:
  schemas:
...
    KundliInput:
      type: object
      required: [name, dateOfBirth, timeOfBirth, placeOfBirth]
      properties:
        name: { type: string }
        dateOfBirth: { type: string }
        timeOfBirth: { type: string }
        placeOfBirth: { type: string }
        gender: { type: string }
        language: { type: string }
    KundliReport:
      type: object
      required: [id, name, dateOfBirth, timeOfBirth, placeOfBirth, sunSign, moonSign, ascendant, analysis, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        dateOfBirth: { type: string }
        timeOfBirth: { type: string }
        placeOfBirth: { type: string }
        sunSign: { type: string }
        moonSign: { type: string }
        ascendant: { type: string }
        analysis: { type: string }
        planetaryPositions: { type: ["string", "null"] }
        doshas: { type: ["string", "null"] }
        remedies: { type: ["string", "null"] }
        manglikStatus: { type: ["string", "null"] }
        luckyStone: { type: ["string", "null"] }
        language: { type: string }
        createdAt: { type: string }

    # PROBLEM SOLVER
...
    VastuInput:
      type: object
      required: [imageBase64, roomType]
      properties:
        imageBase64: { type: string }
        roomType: { type: string, enum: [bedroom, kitchen, living_room, bathroom, office, entrance, garden, full_house] }
        language: { type: string }
    VastuDosh:
      type: object
      required: [name, severity, description, upaay]
      properties:
        name: { type: string }
        severity: { type: string, enum: [high, medium, low] }
        description: { type: string }
        upaay: { type: string }
    VastuReport:
      type: object
      required: [id, roomType, overallScore, findings, remedies, createdAt]
      properties:
        id: { type: integer }
        roomType: { type: string }
        overallScore: { type: integer }
        findings: { type: string }
        remedies: { type: string }
        doshas:
          type: array
          items:
            $ref: "#/components/schemas/VastuDosh"
        positiveAspects: { type: ["string", "null"] }
        negativeAspects: { type: ["string", "null"] }
        colorRecommendations: { type: ["string", "null"] }
        directionAnalysis: { type: ["string", "null"] }
        language: { type: string }
        createdAt: { type: string }

    # IMAGE ANALYSIS (shared for palm & face)
    ImageAnalysisInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64: { type: string }
        language: { type: string }

    # PALM READING
    PalmReport:
      type: object
      required: [id, lifeLine, heartLine, headLine, analysis, createdAt]
      properties:
        id: { type: integer }
        lifeLine: { type: string }
        heartLine: { type: string }
        headLine: { type: string }
        fateLine: { type: ["string", "null"] }
        sunLine: { type: ["string", "null"] }
        analysis: { type: string }
        longevityPrediction: { type: ["string", "null"] }
        careerPrediction: { type: ["string", "null"] }
        lovePrediction: { type: ["string", "null"] }
        fortunePrediction: { type: ["string", "null"] }
        language: { type: string }
        createdAt: { type: string }

    # FACE READING
    FaceReport:
      type: object
      required: [id, faceShape, eyeAnalysis, noseAnalysis, analysis, createdAt]
      properties:
        id: { type: integer }
        faceShape: { type: string }
        eyeAnalysis: { type: string }
        noseAnalysis: { type: string }
        lipsAnalysis: { type: ["string", "null"] }
        foreheadAnalysis: { type: ["string", "null"] }
        analysis: { type: string }
        personalityTraits: { type: ["string", "null"] }
        fortunePrediction: { type: ["string", "null"] }
        healthIndicators: { type: ["string", "null"] }
        language: { type: string }
        createdAt: { type: string }

    # NUMEROLOGY
...